feat: add support of command args with spaces
